# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready Go Fiber boilerplate with RBAC, Redis caching, and PostgreSQL.",
        "title": "Go Fiber Boilerplate API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/api/v1",
    "paths": {
...
        "/rooms": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all rooms with optional filters and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all rooms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room type filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Floor filter",
                        "name": "floor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Building filter",
                        "name": "building",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Active status filter",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/docs.RoomDocResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/docs.MetaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new room in the hospital",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "parameters": [
                    {
                        "description": "Room data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_room.CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.RoomDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a room by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a room by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.RoomDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a room by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Room update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_room.UpdateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.RoomDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a room by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ErrorResponse": {
            "description": "Standard error response wrapper",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error_code": {
                    "type": "integer",
                    "example": -1001
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error occurred"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "docs.MetaResponse": {
            "description": "Pagination metadata for list responses",
            "type": "object",
            "properties": {
                "is_back": {
                    "type": "boolean",
                    "example": false
                },
                "is_next": {
                    "type": "boolean",
                    "example": true
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "docs.RoomDocResponse": {
            "description": "Room information",
            "type": "object",
            "properties": {
                "building": {
                    "type": "string",
                    "example": "Main"
                },
                "capacity": {
                    "type": "integer",
                    "example": 2
                },
                "created_at": {
                    "type": "string"
                },
                "floor": {
                    "type": "string",
                    "example": "1"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Room 101"
                },
                "type": {
                    "type": "string",
                    "example": "patient_room"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "docs.SuccessResponse": {
            "description": "Standard success response wrapper",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "meta": {
                    "$ref": "#/definitions/docs.MetaResponse"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_module_room.CreateRoomRequest": {
            "type": "object",
            "required": [
                "floor",
                "name",
                "type"
            ],
            "properties": {
                "building": {
                    "type": "string",
                    "maxLength": 100
                },
                "capacity": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 0
                },
                "floor": {
                    "type": "string",
                    "maxLength": 20
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "type": {
                    "enum": [
                        "patient_room",
                        "nurse_station",
                        "icu",
                        "emergency",
                        "operating_room"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_module_room.RoomType"
                        }
                    ]
                }
            }
        },
        "internal_module_room.RoomType": {
            "type": "string",
            "enum": [
                "patient_room",
                "nurse_station",
                "icu",
                "emergency",
                "operating_room"
            ],
            "x-enum-varnames": [
                "RoomTypePatientRoom",
                "RoomTypeNurseStation",
                "RoomTypeICU",
                "RoomTypeEmergency",
                "RoomTypeOperatingRoom"
            ]
        },
        "internal_module_room.UpdateRoomRequest": {
            "type": "object",
            "properties": {
                "building": {
                    "type": "string",
                    "maxLength": 100
                },
                "capacity": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 0
                },
                "floor": {
                    "type": "string",
                    "maxLength": 20
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "type": {
                    "enum": [
                        "patient_room",
                        "nurse_station",
                        "icu",
                        "emergency",
                        "operating_room"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_module_room.RoomType"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}. Get token from /auth/login endpoint.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}